+++ src/private/release/InvokeNovaReleaseWorkflow.ps1
@@ -17,6 +17,16 @@ function Get-NovaReleaseNestedWorkflowParameterMap {
     return $nestedWorkflowParams
 }
 
+function Get-NovaReleaseBuildWorkflowParameterMap {
+    [CmdletBinding()]
+    param(
+        [hashtable]$WorkflowParams = @{},
+        [switch]$OverrideWarningRequested
+    )
+
+    return Get-NovaBuildCommandParameterMap -WorkflowParams $WorkflowParams -OverrideWarningRequested:$OverrideWarningRequested
+}
+
 function Test-NovaReleaseWorkflowShouldRestoreBuiltModule {
     [CmdletBinding()]
     param(
@@ -35,18 +45,21 @@ function Invoke-NovaReleaseWorkflow {
 
     $continuousIntegrationRequested = ($WorkflowContext.PSObject.Properties.Name -contains 'ContinuousIntegrationRequested') -and $WorkflowContext.ContinuousIntegrationRequested
     $skipTestsRequested = ($WorkflowContext.PSObject.Properties.Name -contains 'SkipTestsRequested') -and $WorkflowContext.SkipTestsRequested
+    $overrideWarningRequested = ($WorkflowContext.PSObject.Properties.Name -contains 'OverrideWarningRequested') -and $WorkflowContext.OverrideWarningRequested
     $workflowParams = $WorkflowContext.WorkflowParams
     $ciWorkflowParams = Get-NovaReleaseNestedWorkflowParameterMap -WorkflowParams $workflowParams -ContinuousIntegrationRequested:$continuousIntegrationRequested
+    $buildWorkflowParams = Get-NovaReleaseBuildWorkflowParameterMap -WorkflowParams $ciWorkflowParams -OverrideWarningRequested:$overrideWarningRequested
+    $testWorkflowParams = Get-NovaReleaseBuildWorkflowParameterMap -WorkflowParams $workflowParams -OverrideWarningRequested:$overrideWarningRequested
     $shouldRestoreBuiltModule = Test-NovaReleaseWorkflowShouldRestoreBuiltModule -WorkflowParams $workflowParams -ContinuousIntegrationRequested:$continuousIntegrationRequested
     $publishParams = $WorkflowContext.PublishParams
 
-    Invoke-NovaBuild @ciWorkflowParams
+    Invoke-NovaBuild @buildWorkflowParams
     if (-not $skipTestsRequested) {
-        Test-NovaBuild @workflowParams
+        Test-NovaBuild @testWorkflowParams
     }
 
     $versionResult = Update-NovaModuleVersion @ciWorkflowParams
-    Invoke-NovaBuild @ciWorkflowParams
+    Invoke-NovaBuild @buildWorkflowParams
 
     & $WorkflowContext.PublishInvocation.Action @publishParams